requirements.txt
# FuzzMaster Requirements
# Core dependencies for web fuzzing automation

# HTTP libraries
requests>=2.28.0
urllib3>=1.26.0
aiohttp>=3.8.0

# Data processing
pyyaml>=6.0
jsonschema>=4.0.0

python-Levenshtein>=0.12.0

# Progress and CLI
tqdm>=4.64.0
colorama>=0.4.4
click>=8.0.0

# Networking
dnspython>=2.2.0
ipaddress>=1.0.23

# Security
cryptography>=3.4.8
certifi>=2022.0.0

# Configuration
configparser>=5.2.0
python-dotenv>=0.19.0

# Data formats
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Utility
pathlib2>=2.3.0
python-dateutil>=2.8.0

# Optional dependencies for enhanced features
# Machine learning (optional)
# scikit-learn>=1.0.0
# numpy>=1.20.0
# pandas>=1.3.0

# Visualization (optional)
# matplotlib>=3.3.0
# seaborn>=0.11.0

# Database support (optional)
# sqlalchemy>=1.4.0
# sqlite3

# Advanced HTTP features (optional)
# httpx>=0.23.0
# h2>=4.1.0
